+++ src/App.jsx
@@ -6,11 +6,11 @@ import {
   Link,
 } from 'react-router-dom';
 
-import HomePage from './HomePage';
-import AboutPage from './AboutPage';
-import RestaurantsPage from './RestaurantsPage';
-import RestaurantPage from './RestaurantPage';
-import NotFoundPage from './NotFoundPage';
+import HomePage from './pages/HomePage';
+import AboutPage from './pages/AboutPage';
+import RestaurantsPage from './pages/RestaurantsPage';
+import RestaurantPage from './pages/RestaurantPage';
+import NotFoundPage from './pages/NotFoundPage';
 
 // 로그인 컴포넌트 구현
 // 로그인 컨테이너 구현
@@ -24,7 +24,7 @@ import NotFoundPage from './NotFoundPage';
 // 로그인 페이지 구현
 // 레스토랑 리뷰 컨테이너 구현
 // 레스토랑 리뷰 서비스 구현
-/* 
+/*
  *        [POST] https://eatgo-customer-api.ahastudio.com/restaurants/${restaurantId}/reviews
  *        Content-Type : application/json
  *        Authorization : Bearer ${key}
