+++ src/containers/CategoriesContainer.test.jsx
