+++ src/redux/actions.js
@@ -3,7 +3,7 @@ import {
   fetchCategories,
   fetchRestaurants,
   fetchRestaurant,
-} from './services/api';
+} from '../services/api';
 
 export function setRegions(regions) {
   return {
